# docs/swagger.yaml
basePath: /
definitions:
...
  aiovms_internal_model.RecordSchedule:
    properties:
      camera_id:
        type: string
      created_at:
        type: string
      enabled:
        type: boolean
      end_time:
        description: '"20:00"'
        type: string
      id:
        type: string
      last_action:
        description: start / stop / idle
        type: string
      last_triggered_at:
        type: string
      license_id:
        type: integer
      name:
        type: string
      start_time:
        description: '"08:00"'
        type: string
      stream_type:
        description: main / sub
        type: string
      updated_at:
        type: string
      weekdays:
        description: '"1,2,3,4,5" (Sun=0)'
        type: string
    type: object
  aiovms_internal_model.Recording:
    properties:
      camera_id:
        type: string
      codec:
        type: string
      created_at:
        type: string
      duration:
        type: integer
      end_time:
        type: string
      file_path:
        type: string
      file_size:
        type: integer
      filename:
        type: string
      id:
        type: string
      license_id:
        type: integer
      record_type:
        description: manual / scheduled / motion
        type: string
      resolution:
        type: string
      start_time:
        type: string
      status:
        description: recording / complete / truncated
        type: string
      updated_at:
        type: string
    type: object
...
  aiovms_pkg_utils.PaginatedResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/aiovms_pkg_utils.pageData'
      message:
        type: string
    type: object
  aiovms_pkg_utils.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
  aiovms_pkg_utils.pageData:
    properties:
      list: {}
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
...
host: localhost:8081
info:
  contact:
    name: API Support
    url: https://github.com/qoderwork/aiovms
  description: |-
    Video Management Service for AIO NMS.
    Provides camera management, recording, and scheduling APIs. Internal service invoked by NMS proxy.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: AIO VMS API
  version: "1.0"
paths:
...
  /recordings:
    get:
      description: 按摄像头 + 时间范围分页查询录像
      parameters:
      - description: 摄像头 ID
        in: query
        name: camera_id
        type: string
      - description: 开始时间下限（RFC3339）
        in: query
        name: start_time
        type: string
      - description: 开始时间上限（RFC3339）
        in: query
        name: end_time
        type: string
      - default: 1
        description: 页码（从 1 开始）
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/aiovms_pkg_utils.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/aiovms_internal_model.Recording'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/aiovms_pkg_utils.Response'
      security:
      - TenantHeader: []
      - UserHeader: []
      summary: 录像列表（分页）
      tags:
      - 录像管理
  /recordings/{id}:
    delete:
      description: 删除录像文件和数据库记录
      parameters:
      - description: 录像 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/aiovms_pkg_utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/aiovms_pkg_utils.Response'
      security:
      - TenantHeader: []
      - UserHeader: []
      summary: 删除录像
      tags:
      - 录像管理
    get:
      description: 按 ID 查询录像，返回录像信息和回放 URL
      parameters:
      - description: 录像 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/aiovms_pkg_utils.Response'
            - properties:
                data:
                  properties:
                    play_url:
                      type: string
                    recording:
                      $ref: '#/definitions/aiovms_internal_model.Recording'
                  type: object
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/aiovms_pkg_utils.Response'
      security:
      - TenantHeader: []
      - UserHeader: []
      summary: 录像详情
      tags:
      - 录像管理
  /recordings/cameras/{id}/start:
    post:
      description: 对指定摄像头立即开始手动录像
      parameters:
      - description: 摄像头 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/aiovms_pkg_utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/aiovms_pkg_utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/aiovms_pkg_utils.Response'
      security:
      - TenantHeader: []
      - UserHeader: []
      summary: 手动开始录像
      tags:
      - 录像管理
  /recordings/cameras/{id}/stop:
    post:
      description: 对指定摄像头立即停止手动录像
      parameters:
      - description: 摄像头 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/aiovms_pkg_utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/aiovms_pkg_utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/aiovms_pkg_utils.Response'
      security:
      - TenantHeader: []
      - UserHeader: []
      summary: 手动停止录像
      tags:
      - 录像管理
  /schedules:
    get:
      description: 按租户查询录像计划，可选按摄像头过滤
      parameters:
      - description: 摄像头 ID
        in: query
        name: camera_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/aiovms_pkg_utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/aiovms_internal_model.RecordSchedule'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/aiovms_pkg_utils.Response'
      security:
      - TenantHeader: []
      - UserHeader: []
      summary: 录像计划列表
      tags:
      - 录像计划
    post:
      consumes:
      - application/json
      description: 为摄像头创建周期性录像计划
      parameters:
      - description: 录像计划参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/aiovms_internal_model.RecordSchedule'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/aiovms_pkg_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/aiovms_internal_model.RecordSchedule'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/aiovms_pkg_utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/aiovms_pkg_utils.Response'
      security:
      - TenantHeader: []
      - UserHeader: []
      summary: 创建录像计划
      tags:
      - 录像计划
  /schedules/{id}:
    delete:
      description: 删除指定录像计划
      parameters:
      - description: 计划 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/aiovms_pkg_utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/aiovms_pkg_utils.Response'
      security:
      - TenantHeader: []
      - UserHeader: []
      summary: 删除录像计划
      tags:
      - 录像计划
    put:
      consumes:
      - application/json
      description: 更新指定录像计划的参数
      parameters:
      - description: 计划 ID
        in: path
        name: id
        required: true
        type: string
      - description: 录像计划参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/aiovms_internal_model.RecordSchedule'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/aiovms_pkg_utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/aiovms_pkg_utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/aiovms_pkg_utils.Response'
      security:
      - TenantHeader: []
      - UserHeader: []
      summary: 更新录像计划
      tags:
      - 录像计划
  /schedules/{id}/toggle:
    patch:
      description: 翻转录像计划的 enabled 标志
      parameters:
      - description: 计划 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/aiovms_pkg_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/aiovms_internal_model.RecordSchedule'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/aiovms_pkg_utils.Response'
      security:
      - TenantHeader: []
      - UserHeader: []
      summary: 启停录像计划
      tags:
      - 录像计划
securityDefinitions:
  TenantHeader:
    description: Tenant (license) ID, forwarded by NMS. Use "default" for tenant 1.
    in: header
    name: X-License-Id
    type: apiKey
  UserHeader:
    description: User ID, forwarded by NMS.
    in: header
    name: X-User-Id
    type: apiKey
swagger: "2.0"